requirements.txt
streamlit
pandas
numpy
plotly
google-cloud-bigquery
db-dtypes
requests
